refactor: move tunnelTokenGate.ts to src/vite-middleware
Keeps the dev-server middleware under src/ so the default tsconfig include
covers it, instead of listing the file at the repo root explicitly.
