+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.js", "styled-system", "tunnelTokenGate.ts"],
+  "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.js", "styled-system"],
   "compilerOptions": {
     "allowJs": true,
     "allowSyntheticDefaultImports": true,
